docs+test(client): address PR #455 third review (migration, ARCH, trybuild)
- Add 0.18 migration guide for registry/FeatureLoader contraction
- Update ARCHITECTURE ServiceRegistry sections to metadata-only truth
- trybuild compile-fail for generation-time catalog uniqueness (#423)
