+++ crates/openlark-client/Cargo.toml
@@ -91,3 +91,4 @@ wiremock = { workspace = true }
 rstest = { workspace = true }
 tempfile = { workspace = true }
 mockall = { workspace = true }
+trybuild = "1.0"
