+++ crates/openlark-client/tests/ui.rs
@@ -1,0 +1,8 @@
+//! compile-fail 回归：capability catalog generation-time 唯一性（#423 / #455）
+
+#[test]
+fn capability_catalog_generation_time_uniqueness() {
+    let t = trybuild::TestCases::new();
+    t.compile_fail("tests/ui/duplicate_capability_field.rs");
+    t.compile_fail("tests/ui/mismatch_capability_name.rs");
+}
